initial push
These are the first draft schemas.  The appDef schema defines the configuration for a specific application while buildDef defines the steps involved in generating an app from and appDef.  componentDef and uiDef are generic schemas for defining the ui layer of the app, but may be replaced by a more mature model such as Storybook.js